Sailors refactoring

+++ src/libs/sailors/src/sailors_way_points.h
@@ -150,8 +150,8 @@ class SailorsPoints
 
 inline float Dest(const CVECTOR &_v1, const CVECTOR &_v2)
 {
-    return static_cast<float>(sqrt((_v2.x - _v1.x) * (_v2.x - _v1.x) + (_v2.y - _v1.y) * (_v2.y - _v1.y) +
-                                   (_v2.z - _v1.z) * (_v2.z - _v1.z)));
+    return sqrt((_v2.x - _v1.x) * (_v2.x - _v1.x) + (_v2.y - _v1.y) * (_v2.y - _v1.y) +
+                (_v2.z - _v1.z) * (_v2.z - _v1.z));
 };
 
 inline bool Dest(const CVECTOR &_v1, const CVECTOR &_v2, float d)
@@ -161,7 +161,7 @@ inline bool Dest(const CVECTOR &_v1, const CVECTOR &_v2, float d)
 
 inline float Vector2Angle(const CVECTOR &_v)
 {
-    auto result = static_cast<float>(atan2(_v.x, _v.z));
+    auto result = atan2(_v.x, _v.z);
 
     while (result >= PI * 2)
         result -= PI * 2;
